requirements/local.txt
amqp==5.3.1
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
black==24.8.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.39.1
cron_descriptor==2.0.6
cryptography==46.0.1
defusedxml==0.7.1
django
Django==5.0.14
django-appconf==1.1.0
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-countries==7.6.1
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
djoser==2.2.3
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
loguru==0.7.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
phonenumbers==8.13.42
pillow==10.4.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.14



